wip: pre release enhancements
