+++ examples/basic/basic_ref.html
@@ -1,0 +1,3 @@
+<script src="./basic_ref.js" type="module"></script>
+
+<my-app></my-app>
