+++ examples/dom-helpers/dom-helpers_ref.js
@@ -36,7 +36,6 @@ let app = create({
       e.target.innerHTML = '';
     },
   },
-
   children: [
     {
       processors: [
